Contact.java: Extract commonalities from Phone, Email and Address Address.java, Email.java, Phone.java: Remove the commonalities
